Make the tree trimming an order of magnitude faster. The revised process avoids repeated cache deletion passes over the tree - these were going quadratic and thus very slow for large trees. It is also more efficient in determining which nodes to keep (if a node is kept then so too are its ancestors). Also do a bit of code clean up and conversion of while loops to foreach,  These avoid possible hash problems and should be more efficient since we more often need the node ref than its name, and we can get the name as needed.

+++ lib/Biodiverse/Tree.pm
@@ -140,6 +140,8 @@ sub set_parents_below {
     return;
 }
 
+#  If no_delete_cache is true then the caller promises to clean up later.
+#  This can be used to avoid multiple passes over the tree across multiple deletions.  
 sub delete_node {
     my $self = shift;
     my %args = @_;
@@ -148,28 +150,25 @@ sub delete_node {
     my $node_ref = $self->get_node_ref (node => $args{node});
     return if ! defined $node_ref;  #  node does not exist anyway
 
-    #  Clear any cached values.
-    #  We could just do the DESCENDENTS lists, but they are not the
-    #  only ones that point to now non-existent nodes.
-    #  This also circumvents circular refs from the caches.
-    #  ...->these are done later in the sub now to reduce multiple cache deletion passes.
-    #$self->get_tree_ref->delete_cached_values_below;
-
     #  get the names of all descendents 
     my %node_hash = $node_ref->get_all_descendents (cache => 0);
     $node_hash{$node_ref->get_name} = $node_ref;  #  add node_ref to this list
 
-    #  Now we delete it from the treenode structure.  This cleans up any children in the tree.
+    #  Now we delete it from the treenode structure.
+    #  This cleans up any children in the tree.
     $node_ref->get_parent->delete_child (child => $node_ref, no_delete_cache => 1);
 
     #  now we delete it and its descendents from the node hash
     $self->delete_from_node_hash (nodes => \%node_hash);
 
-    #  now we clear the caches from those deleted nodes and those remaining
-    foreach my $n_ref (values %node_hash) {
-        $n_ref->delete_cached_values;
+    #  Now we clear the caches from those deleted nodes and those remaining
+    #  This circumvents circular refs from the caches.
+    if (!$args{no_delete_cache}) {
+        foreach my $n_ref (values %node_hash) {
+            $n_ref->delete_cached_values;
+        }
+        $self->delete_cached_values_below;
     }
-    $self->delete_cached_values_below;
 
     #  return a list of the names of those deleted nodes
     return wantarray ? keys %node_hash : [keys %node_hash];
@@ -459,7 +458,7 @@ sub node_is_in_tree {
 
     #  node cannot exist if it has no name...
     croak "node name undefined\n"
-      if !defined $node_name;  
+      if !defined $node_name;
 
     my $node_hash = $self->get_node_hash;
     return exists $node_hash->{$node_name};
@@ -469,9 +468,9 @@ sub get_terminal_nodes {
     my $self = shift;
     my %node_list;
 
-    while ((my $node, my $node_ref) = each (%{$self->get_node_hash})) {
+    foreach my $node_ref (values %{$self->get_node_hash}) {
         next if ! $node_ref->is_terminal_node;
-        $node_list{$node} = $node_ref;
+        $node_list{$node_ref->get_name} = $node_ref;
     }
 
     return wantarray ? %node_list : \%node_list;
@@ -481,7 +480,7 @@ sub get_terminal_node_refs {
     my $self = shift;
     my @node_list;
 
-    while ((my $node, my $node_ref) = each (%{$self->get_node_hash})) {
+    foreach my $node_ref (values %{$self->get_node_hash}) {
         next if ! $node_ref->is_terminal_node;
         push @node_list, $node_ref;
     }
@@ -494,11 +493,11 @@ sub get_root_nodes {  #  if there are several root nodes
     my %node_list;
     my $node_hash = $self->get_node_hash;
 
-    while ((my $node, my $node_ref) = each (%$node_hash)) {
-        next if (! defined $node_ref);
-        $node_list{$node} = $node_ref if $node_ref->is_root_node;
-        #my $check = $node_ref->is_root_node;
-        #print "";
+    foreach my $node_ref (values %$node_hash) {
+        next if ! defined $node_ref;
+        if ($node_ref->is_root_node) {
+            $node_list{$node_ref->get_name} = $node_ref ;
+        }
     }
 
     return wantarray ? %node_list : \%node_list;
@@ -529,21 +528,21 @@ sub get_named_nodes {
     my $self = shift;
     my %node_list;
     my $node_hash = $self->get_node_hash;
-    while (my ($node, $node_ref) = each (%$node_hash)) {
+    foreach my $node_ref (values %$node_hash) {
         next if $node_ref->is_internal_node;
-        $node_list{$node} = $node_ref;
+        $node_list{$node_ref->get_name} = $node_ref;
     }
     return wantarray ? %node_list : \%node_list;
 }
 
-#  get all the nodes that aren't named nodes
+#  get all the nodes that aren't terminals
 sub get_branch_nodes {
     my $self = shift;
     my %node_list;
     my $node_hash = $self->get_node_hash;
-    while (my ($node, $node_ref) = each (%$node_hash)) {
+    foreach my $node_ref (values %$node_hash) {
         next if $node_ref->is_terminal_node;
-        $node_list{$node} = $node_ref;
+        $node_list{$node_ref->get_name} = $node_ref;
     }
     return wantarray ? %node_list : \%node_list;
 }
@@ -551,7 +550,7 @@ sub get_branch_nodes {
 sub get_branch_node_refs {
     my $self = shift;
     my @node_list;
-    while ((my $node, my $node_ref) = each (%{$self->get_node_hash})) {
+    foreach my $node_ref (values %{$self->get_node_hash}) {
         next if $node_ref->is_terminal_node;
         push @node_list, $node_ref;
     }
@@ -1796,29 +1795,44 @@ sub trim {
         @_,                      #  if they have no named children to keep
     );
 
-    my $delete_internals = $args{delete_internals};
+    say '[TREE] Trimming';
 
-    #  get keep and trim lists and convert to hashes as needs dictate
-    my $keep = $args{keep} || {};  #  those to keep
-    $keep = $self->array_to_hash_keys (list => $keep);
+    my $delete_internals = $args{delete_internals};
+    
+    my %tree_node_hash = $self->get_node_hash;
+
+    #  Get keep and trim lists and convert to hashes as needs dictate
+    #  those to keep
+    my $keep = $self->array_to_hash_keys (list => $args{keep} || {});
     my $trim = $args{trim}; #  those to delete
 
-    #  if the keep list is defined, and the trim list is not defined,
+    #  If the keep list is defined, and the trim list is not defined,
     #    then we work with all named nodes that don't have children we want to keep
     if (! defined $args{trim} && defined $args{keep}) {
-        $trim = {};
-        my %node_hash = $self->get_node_hash;
-        foreach my $name (keys %node_hash) {
-            next if exists $keep->{$name};
-            my $node = $node_hash{$name};
-            next if $node->is_internal_node;
-            next if $node->is_root_node;  #  never delete the root node
-            my %children = $node->get_all_descendents;  #  make sure we use a copy
+
+      NAME:
+        foreach my $name (keys %tree_node_hash) {
+            next NAME if exists $keep->{$name};
+
+            my $node = $tree_node_hash{$name};
+
+            next NAME if $node->is_internal_node;
+            next NAME if $node->is_root_node;  #  never delete the root node
+
+            my %children    = $node->get_all_descendents;  #  make sure we use a copy
             my $child_count = scalar keys %children;
             delete @children{keys %$keep};
-            #  if any were deleted then we have some children to keep.  don't add this node
-            next if $child_count != scalar keys %children;  
-            $trim->{$name} = $node_hash{$name};
+            #  If none were deleted then we can trim this node.
+            #  Otherwise add this node and all of its ancestors to the keep list.
+            if ($child_count == scalar keys %children) {
+                $trim->{$name} = $node;
+            }
+            else {
+                my $ancestors = $node->get_path_to_root_node;
+                foreach my $ancestor (@$ancestors) {
+                    $keep->{$ancestor->get_name} ++;
+                }
+            }
         }
     }
     else {
@@ -1827,56 +1841,92 @@ sub trim {
     my %trim_hash = $self->array_to_hash_keys (list => $trim);  #  makes a copy
 
     #  we only want to consider those not being explicitly kept (included)
-    my %candidate_node_hash = $self->get_node_hash;
+    my %candidate_node_hash = %tree_node_hash;
     delete @candidate_node_hash{keys %$keep};
 
     my %deleted_h;
+    my $i = 0;
+    my $to_do = scalar keys %candidate_node_hash;
+    my $progress = Biodiverse::Progress->new (text => 'Deletions');
 
   DELETION:
     foreach my $name (keys %candidate_node_hash) {
+        $i++;
         #  we might have deleted a named parent,
         #  so this node no longer exists in the tree
         next DELETION if $deleted_h{$name} || !exists $trim_hash{$name};
 
+        $progress->update (
+            "Deleting nodes ($i / $to_do)",
+            $i / $to_do,
+        );
+
         #  delete if it is in the list to exclude
-        say "[Tree] Deleting node $name";
-        my @deleted_nodes = $self->delete_node (node => $name);
+        my @deleted_nodes = $self->delete_node (node => $name, no_delete_cache => 1);
         @deleted_h{@deleted_nodes} = (1) x scalar @deleted_nodes;
-        #foreach my $del_name (@deleted_nodes) {
-        #    $deleted_h{$del_name} ++;
-        #}
     }
 
+    $progress->close_off;
+    my $deleted_count = scalar keys %deleted_h;
+    say "[TREE] Deleted $deleted_count nodes ", join ' ', sort keys %deleted_h;
+
+    #  delete any internal nodes with no named descendents
+    my $deleted_internal_count = 0;
     if ($delete_internals and scalar keys %deleted_h) {
-        #  delete any internal nodes with no named children
+        say '[TREE] Cleaning up internal nodes';
+
         my %node_hash = $self->get_node_hash;
+        $to_do = scalar keys %node_hash;
+        my %deleted_hash;
+        my $i;
+
       NODE:
         foreach my $name (keys %node_hash) {
+            $i++;
             my $node = $node_hash{$name};
-            next NODE if !$node->is_internal_node || $node->is_root_node;
+            next NODE if $deleted_hash{$node};  #  already deleted
+            next NODE if !$node->is_internal_node;
+            next NODE if  $node->is_root_node;
+
+            $progress->update (
+                "Deleting nodes ($i / $to_do)",
+                $i / $to_do,
+            );
 
             my $children = $node->get_all_descendents;
-            my %named_children;
+          DESCENDENT:
             foreach my $child (keys %$children) {
                 my $child_node = $children->{$child};
-                $named_children{$child} = 1 if ! $child_node->is_internal_node;
+                next NODE if ! $child_node->is_internal_node;
             }
-            if (scalar keys %named_children == 0) {
-                #  might have already been deleted , so wrap in an eval
-                eval { $self->delete_node (node => $name) };
-            }
+            #  might have already been deleted, so wrap in an eval
+            my @deleted_names = eval {
+                $self->delete_node (node => $name, no_delete_cache => 1)
+            };
+            @deleted_hash{@deleted_names} = (1) x @deleted_names;
         }
+        $progress->close_off;
+
+        $deleted_internal_count = scalar keys %deleted_hash;
+        say "[TREE] Deleted $deleted_internal_count internal nodes with no named descendents";
     }
 
-    $self->delete_param ('TOTAL_LENGTH');  #  need to clear this up
-    $self->get_tree_ref->delete_cached_values_below;
+    #  now some cleanup
+    if ($deleted_internal_count || $deleted_count) {
+        $self->delete_param ('TOTAL_LENGTH');  #  need to clear this up
+        
+        #  This avoids circular refs in the ones that were deleted
+        foreach my $node (values %tree_node_hash) {
+            $node->delete_cached_values;
+        }
+    }
     $keep = undef;  #  was leaking - not sure it matters, though
+
+    say '[TREE] Trimming completed';
 
     return $self;
 }
 
-#sub min {return $_[0] > $_[1] ? $_[1] : $_[0]};
-#sub max {return $_[0] < $_[1] ? $_[1] : $_[0]};
 
 sub numerically {$a <=> $b};
